updating zip code level map
